Add ycrc
YCql Row Count program.

This provides a compiled binary to execute a count(*) across partitions
in a CQL cluster.

The default scaling factor of 6 and default parallel tasks of 16 should
be performant for most use cases, but the values can be scaled using
--scale and --parallel flags respectively.

Of note, in order to avoid sequential calls to a single partition, the
partition range map is randomized before queries begin. While not
strictly guaranteed to hit different tablets, I hope this is sufficient
to avoid hammering a single tablet while others are untouched.

TLS and authentication are available. root CA cert and user/password
authentication is tested - as of this commit, key/cert is not tested,
but it's directly passed to the driver, so it's unlikely that this code
causes issues, if the driver is sound.

Signed-off-by: Tyler Ramer <tramer@yugabyte.com>