yugatool: attempt cleartext connection if TLS connection fails
Partial fix for https://github.com/yugabyte/yb-tools/issues/28

This commit implements retry in cleartext if TLS connectivity fails but not vice-versa.

A complete fix will require a bit more refactoring since we have to keep track of whether we've already tried to swap from TLS to cleartext or vice-versa, otherwise we will end up in a loop.