rotatecert.sh: 0.1.2 - Fix Subject Key Identifier
Previous versions of the script depended on the openssl -ext flag which is not available in ancient (1.0.2) versions of openssl. The main change here is to move to grep for detecting the Subject Key Identifier. Screen scraping is not ideal but the blast radius of this particular grep is pretty small.

There are also a handful of smaller changes:
- Set bash configs to abort on unbound variables and pipe failures
- Add missing error handling for certificate generation failures
- Formatting consistency improvements in the OpenSSL cnf files