rotatecert.sh: 0.1.3 - More Subject Key Identifier fixes
The primary fix in this commit corrects an issue where the previous fix to Subject Key Identifier did not differentiate between a failure of the openssl command and a failure of the subsequent grep.

This commit splits the check into two separate operations, so if the openssl check fails, that is fatal but if the grep fails, we treat that as a missing Subject Key Identifier. This should address the issue.

Other minor changes:
- Added a header with instructions on how to use the script
- Updated ROOT_CERT_DIR variable to support setting at the CLI with export
- Fixed typo in TODO comment