rotatecert.sh: 0.1.4 - Logging and resiliency improvements
The primary change in this release is that the script now generates a log file instead of relying on the user to redirect the output somewhere.

To support this, the various debug, info, warn, etc. functions have been refactored so they call a common log_message function that has the logic for whether to append to the log or just log to the terminal

Added a signal handler to ensure that we log something if the script is aborted with ^C.

If something goes wrong while creating the log file, file logging will be disabled with a warning.

Variablized the path to the mkdir binary.

Improved handling for incorrect certificate and SSH key paths.

Added logging for the local platform certificate backup operation.